refactor: move point filter state and data fetching down to container
기존에는 MyPoints 컴포넌트에서 필터 타입(type) 상태와 useUserPointHistory 훅을 함께 관리하고 있었지만,
해당 데이터가 실제로는 PointHistoryContainer에서만 사용되므로, 관심사 분리 및 불필요한 리렌더링을 피하기 위해 로직을 이동시켰습니다.
